Added API to update gameState with each button push (next step, figure out how to process changed gameState)
